+++ lib/src/config.dart
@@ -5,6 +5,7 @@ class UnleashConfig {
     required this.proxyUrl,
     required this.clientKey,
     this.instanceId,
+    this.poolMode = const Duration(seconds: 15),
   });
 
   /// Unleash Proxy URL
@@ -15,6 +16,9 @@ class UnleashConfig {
 
   /// Unleash Instance Id
   final String? instanceId;
+
+  /// Unleash Polling Mode
+  final Duration poolMode;
 
   /// Header default for Unleash Request
   Map<String, String> get headers => {