+++ lib/src/platform/unleash_platform.dart
@@ -1,0 +1,5 @@
+/// Abstract class for Unleash Platform
+abstract class UnleashPlatform {
+  /// Unleash Platform constructor
+  UnleashPlatform();
+}