+++ lib/src/polling/polling.dart
@@ -1,0 +1,8 @@
+/// Unleash Polling Mode
+class UnleashPollingMode {
+  /// Polling is disabled
+  static Duration none = Duration.zero;
+
+  /// Custom polling interval
+  static Duration custom(Duration duration) => duration;
+}