+++ lib/unleash.dart
@@ -8,7 +8,22 @@
 /// A Very Good Project created by Very Good CLI.
 library unleash_flutter_proxy_sdk;
 
+import 'dart:async';
+
+import 'package:collection/collection.dart';
+import 'package:unleash_flutter_proxy_sdk/src/client.dart';
+import 'package:unleash_flutter_proxy_sdk/src/config.dart';
+import 'package:unleash_flutter_proxy_sdk/src/context.dart';
+import 'package:unleash_flutter_proxy_sdk/src/data/toggle.dart';
+import 'package:unleash_flutter_proxy_sdk/src/platform/unleash_app_platform.dart';
+import 'package:unleash_flutter_proxy_sdk/src/platform/unleash_platform.dart';
+import 'package:unleash_flutter_proxy_sdk/src/polling/polling.dart';
+import 'package:unleash_flutter_proxy_sdk/src/utils.dart';
+
 export 'src/client.dart';
 export 'src/config.dart';
 export 'src/context.dart';
-export 'src/unleash.dart';
+export 'src/polling/polling.dart';
+
+part 'src/unleash.dart';
+part 'src/unleash_app.dart';