+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.1.0+1
 
-- feat: initial commit 🎉
+- feat: get toggle status
+- feat: pooling update config